Replace float timeout by int timeout

+++ render_service_scripts/configure_viewer.py
@@ -118,7 +118,7 @@ def main():
 	if os.path.isfile('RadeonProViewer.exe'):
 		p = psutil.Popen('RadeonProViewer.exe', stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 		try:
-			stdout, stderr = p.communicate(timeout=float(args.timeout))
+			stdout, stderr = p.communicate(timeout=int(args.timeout))
 		except (subprocess.TimeoutExpired, psutil.TimeoutExpired) as err:
 			try:
 				for child in reversed(p.children(recursive=True)):
